requirements.txt
Flask==2.0.1
requests==2.26.0
Pillow==8.4.0
